Run benchmark in infinite loop

+++ caffe2/share/contrib/benchmark_harness/git_driver.py
@@ -2,6 +2,7 @@
 
 import os
 import shutil
+import time
 from utils.arg_parse import getParser, getArgs, parse
 from utils.git import Git
 from utils.custom_logger import getLogger
@@ -21,6 +22,8 @@ getParser().add_argument("--android", action="store_true",
     help="Run the benchmark on all collected android devices.")
 getParser().add_argument("--local_reporter",
     help="Save the result to a directory specified by this argument.")
+getParser().add_argument("--interval", type=int,
+    help="The minimum time interval in seconds between two benchmark runs.")
 
 class GitDriver(object):
     def __init__(self):
@@ -32,7 +35,9 @@ class GitDriver(object):
         if getArgs().git_commit:
             self.git.pull("sf", "benchmarking")
             self.git.checkout(getArgs().git_commit)
-            self.commit_hash = self.git.run('rev-parse', 'HEAD')
+            new_commit_hash = self.git.run('rev-parse', 'HEAD')
+            if new_commit_hash == self.commit_hash:
+                return false
             if getArgs().android:
                 shutil.rmtree(getArgs().git_dir + "/build_android")
                 build_android = getArgs().git_dir + "/scripts/build_android.sh"
@@ -41,6 +46,7 @@ class GitDriver(object):
                 shutil.rmtree(getArgs().git_dir + "/build")
                 build_local = getArgs().git_dir + "/scripts/build_local.sh"
                 processRun(build_local)
+            return true
 
     def _processConfig(self):
         dir_path = os.path.dirname(os.path.realpath(__file__))
@@ -56,14 +62,24 @@ class GitDriver(object):
                 for x in configs]
         return configs
 
-    def run(self):
-        self._setupGit()
-        configs = self._processConfig()
+    def runOnce(self, configs):
+        if not self._setupGit():
+            return
         for config in configs:
             cmds = config.split(' ')
             cmd = [x.strip() for x in cmds]
             getLogger().info("Running: %s", ' '.join(cmd))
             processRun(cmd)
+
+    def run(self):
+        configs = self._processConfig()
+        interval = getArgs().interval if getArgs().interval else 300
+        while True:
+            prev_ts = time.time()
+            self.runOnce(configs)
+            current_ts = time.time()
+            if current_ts < prev_ts + interval:
+                time.sleep(prev_ts + interval - current_ts)
 
 if __name__ == "__main__":
     app = GitDriver()
